Long DAO type

+++ src/main/java/com/goit/popov/restaurant/controller/ImageController.java
@@ -26,7 +26,7 @@ public class ImageController {
         private EmployeeService employeeService;
 
         @RequestMapping(value = "/photo", method = RequestMethod.GET)
-        public void showPhoto(@RequestParam("id") int id,
+        public void showPhoto(@RequestParam("id") Long id,
                               HttpServletResponse response, HttpServletRequest request) throws IOException {
                 Employee employee = employeeService.getById(id);
                 response.setContentType("image/jpeg, image/jpg, image/png, image/gif");
