+++ src/main/java/com/goit/popov/restaurant/dao/UnitDAO.java
@@ -1,13 +1,13 @@
 package com.goit.popov.restaurant.dao;
 
 import com.goit.popov.restaurant.model.Unit;
-
 import java.util.List;
 
 /**
  * Created by Andrey on 11/28/2016.
  */
 public interface UnitDAO {
+
         List<Unit> getAll();
-        Unit getById(int id);
+        Unit getById(Long id);
 }
