+++ src/main/java/com/goit/popov/restaurant/service/authentification/Employee.java
@@ -10,19 +10,19 @@ import java.util.Collection;
  */
 public class Employee extends User {
 
-        private int id;
+        private Long id;
 
-        public int getId() {
+        public Long getId() {
                 return id;
         }
 
-        public void setId(int id) {
+        public void setId(Long id) {
                 this.id = id;
         }
 
         public Employee(String username, String password, boolean enabled, boolean accountNonExpired,
                         boolean credentialsNonExpired, boolean accountNonLocked,
-                        Collection<? extends GrantedAuthority> authorities, int id) {
+                        Collection<? extends GrantedAuthority> authorities, Long id) {
                 super(username, password, enabled, accountNonExpired, credentialsNonExpired, accountNonLocked, authorities);
                 setId(id);
         }
